+++ SPA/NoorCanvas/Models/Question.cs
@@ -10,7 +10,7 @@ public class Question
     public long QuestionId { get; set; }
 
     [Required]
-    public long SessionId { get; set; }
+    public int SessionId { get; set; }
 
     [Required]
     public Guid UserId { get; set; }
